backend/requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
python-multipart==0.0.22
pandas==3.0.0
pillow==12.1.1
